+++ src/ResourceManager/Profile/Commands.Profile.Test/ProfileCmdletTests.cs
@@ -49,9 +49,11 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
         {
             var profile = new AzureRMProfile();
             profile.Environments.Add("foo", AzureEnvironment.PublicEnvironments.Values.FirstOrDefault());
-            SelectAzureRMProfileCommand cmdlt = new SelectAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            ImportAzureRMContextCommand cmdlt = new ImportAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             // Setup
-            cmdlt.Profile = profile;
+            cmdlt.AzureContext = profile;
             cmdlt.CommandRuntime = commandRuntimeMock;
 
             // Act
@@ -65,27 +67,15 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void SelectAzureProfileNull()
-        {
-            SelectAzureRMProfileCommand cmdlt = new SelectAzureRMProfileCommand();
-            // Setup
-            cmdlt.CommandRuntime = commandRuntimeMock;
-
-            // Act
-            cmdlt.InvokeBeginProcessing();
-            Assert.Throws<ArgumentException>(() => cmdlt.ExecuteCmdlet());
-            cmdlt.InvokeEndProcessing();
-        }
-
-
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void SelectAzureProfileBadPath()
         {
-            SelectAzureRMProfileCommand cmdlt = new SelectAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            ImportAzureRMContextCommand cmdlt = new ImportAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             cmdlt.Path = "z:\non-existent-path\non-existent-file.ext";
             // Setup
             cmdlt.CommandRuntime = commandRuntimeMock;
+            cmdlt.MyInvocation.BoundParameters.Add("Path", cmdlt.Path);
 
             // Act
             cmdlt.InvokeBeginProcessing();
@@ -99,11 +89,13 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
             var profile = new AzureRMProfile();
             profile.Environments.Add("foo", AzureEnvironment.PublicEnvironments.Values.FirstOrDefault());
             profile.Save("X:\\foo.json");
-            SelectAzureRMProfileCommand cmdlt = new SelectAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            ImportAzureRMContextCommand cmdlt = new ImportAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             // Setup
             cmdlt.Path = "X:\\foo.json";
             cmdlt.CommandRuntime = commandRuntimeMock;
-
+            cmdlt.MyInvocation.BoundParameters.Add("Path", cmdlt.Path);
             // Act
             cmdlt.InvokeBeginProcessing();
             cmdlt.ExecuteCmdlet();
@@ -119,7 +111,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
         {
             var profile = new AzureRMProfile();
             profile.Environments.Add("foo", AzureEnvironment.PublicEnvironments.Values.FirstOrDefault());
-            SaveAzureRMProfileCommand cmdlt = new SaveAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            SaveAzureRMContextCommand cmdlt = new SaveAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             // Setup
             cmdlt.Profile = profile;
             cmdlt.Path = "X:\\foo.json";
@@ -140,7 +134,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void SaveAzureProfileNull()
         {
-            SaveAzureRMProfileCommand cmdlt = new SaveAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            SaveAzureRMContextCommand cmdlt = new SaveAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             // Setup
             AzureRmProfileProvider.Instance.Profile = null;
             cmdlt.Path = "X:\\foo.json";
@@ -158,7 +154,9 @@ namespace Microsoft.Azure.Commands.ResourceManager.Profile.Test
             profile.Environments.Add("foo", AzureEnvironment.PublicEnvironments.Values.FirstOrDefault());
             profile.Context = new AzureContext(new AzureSubscription(), new AzureAccount(), profile.Environments["foo"]);
             AzureRmProfileProvider.Instance.Profile = profile;
-            SaveAzureRMProfileCommand cmdlt = new SaveAzureRMProfileCommand();
+#pragma warning disable CS0618 // Suppress obsolescence warning: cmdlet name is changing
+            SaveAzureRMContextCommand cmdlt = new SaveAzureRMContextCommand();
+#pragma warning restore CS0618 // Suppress obsolescence warning: cmdlet name is changing
             // Setup
             cmdlt.Path = "X:\\foo.json";
             cmdlt.CommandRuntime = commandRuntimeMock;