+++ src/ResourceManager/Profile/Commands.Profile/AzureRmProfileStartup.ps1
@@ -15,4 +15,6 @@
 @{
     "Login-AzureRmAccount" = "Add-AzureRmAccount";
     "Select-AzureRmSubscription" = "Set-AzureRmContext";
+	"Save-AzureRmProfile" = "Save-AzureRmContext";
+	"Select-AzureRmProfile" = "Import-AzureRmContext";
 }.GetEnumerator() | Select @{Name='Name'; Expression={$_.Key}}, @{Name='Value'; Expression={$_.Value}} | New-Alias -Description "AzureAlias"